+++ minigu/storage/src/common/wal/graph_wal.rs
@@ -29,7 +29,7 @@ use serde::{Deserialize, Serialize};
 
 use super::{LogRecord, StorageWal};
 use crate::error::{StorageError, StorageResult, WalError};
-use crate::tp::transaction::{DeltaOp, IsolationLevel, Timestamp};
+use crate::tp_storage::transaction::{DeltaOp, IsolationLevel, Timestamp};
 
 const HEADER_SIZE: usize = 8; // 4 bytes length + 4 bytes crc32
 const WAL_PATH: &str = "/tmp/wal.log";
@@ -370,7 +370,7 @@ mod tests {
     use serial_test::serial;
 
     use super::*;
-    use crate::tp::transaction::{DeltaOp, SetPropsOp, Timestamp};
+    use crate::tp_storage::transaction::{DeltaOp, SetPropsOp, Timestamp};
 
     fn temp_wal_path() -> PathBuf {
         let mut path = std::env::temp_dir();
