+++ minigu/storage/src/tp_storage/iterators/vertex_iterator.rs
@@ -6,9 +6,9 @@ use minigu_common::types::VertexId;
 use crate::common::iterators::{ChunkData, VertexIteratorTrait};
 use crate::common::model::vertex::Vertex;
 use crate::error::StorageResult;
-use crate::tp::iterators::adjacency_iterator::AdjacencyIterator;
-use crate::tp::memory_graph::VersionedVertex;
-use crate::tp::transaction::MemTransaction;
+use crate::tp_storage::iterators::adjacency_iterator::AdjacencyIterator;
+use crate::tp_storage::memory_graph::VersionedVertex;
+use crate::tp_storage::transaction::MemTransaction;
 
 type VertexFilter<'a> = Box<dyn Fn(&Vertex) -> bool + 'a>;
