+++ minigu/storage/src/tp_storage/transaction.rs
@@ -860,7 +860,7 @@ impl Clone for TransactionHandle {
 #[cfg(test)]
 mod tests {
     use super::{IsolationLevel, *};
-    use crate::tp::memory_graph;
+    use crate::tp_storage::memory_graph;
 
     #[test]
     fn test_watermark_tracking() {
